requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
flake8==6.1.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
mccabe==0.7.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
